+++ packages/demo/__stories__/plugin-image.tsx
@@ -6,22 +6,32 @@ import { EditorStory } from '../src'
 storiesOf('Plugins/Image', module)
   .add('Initial State', () => {
     const state = {
-      plugin: 'image'
+      plugin: 'rows',
+      state: [
+        {
+          plugin: 'image'
+        }
+      ]
     }
     return <EditorStory initialState={state} />
   })
   .add('Prefilled', () => {
     const state = {
-      plugin: 'image',
-      state: {
-        src:
-          'https://packages.serlo.org/athene2-assets@a/serlo_learning_lg.ee37b05f.jpg',
-        href: '',
-        target: '',
-        rel: '',
-        description: 'Ein Schüler lernt mit serlo.org',
-        maxWidth: 0
-      }
+      plugin: 'rows',
+      state: [
+        {
+          plugin: 'image',
+          state: {
+            src:
+              'https://packages.serlo.org/athene2-assets@a/serlo_learning_lg.ee37b05f.jpg',
+            href: '',
+            target: '',
+            rel: '',
+            description: 'Ein Schüler lernt mit serlo.org',
+            maxWidth: 0
+          }
+        }
+      ]
     }
     return <EditorStory initialState={state} />
   })