+++ packages/demo/__stories__/plugin-video.tsx
@@ -6,14 +6,24 @@ import { EditorStory } from '../src'
 storiesOf('Plugins/Video', module)
   .add('Initial State', () => {
     const state = {
-      plugin: 'video'
+      plugin: 'rows',
+      state: [
+        {
+          plugin: 'video'
+        }
+      ]
     }
     return <EditorStory initialState={state} />
   })
   .add('Prefilled', () => {
     const state = {
-      plugin: 'video',
-      state: 'https://www.youtube.com/watch?v=SCJ7nzKwnYo'
+      plugin: 'rows',
+      state: [
+        {
+          plugin: 'video',
+          state: 'https://www.youtube.com/watch?v=SCJ7nzKwnYo'
+        }
+      ]
     }
     return <EditorStory initialState={state} />
   })