+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to this project will be documented in this file.
 ### Internal
 
 - **demo**. Log errors only
+- **demo**. Add additional examples for plugin-usage in Storybook
 
 ## [0.4.3](https://github.com/edtr-io/edtr-io/compare/v0.4.2..v0.4.3) - March 29, 2019
 
@@ -189,7 +190,6 @@ All notable changes to this project will be documented in this file.
 - Add `StateType.scalar` that represents a value of the given type (more specifically: `StateType.boolean`, `StateType.number`, `StateType.child`
 - Add `StateType.list`, `State.object`
 - Add slate for rich text and links. Standard Bold + Italic hotkeys supported.
-- Add additional examples for plugin-usage in Storybook
 
 ### Changed